backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
psutil==6.1.1
pyserial==3.5
gpsd-py3==0.3.0
aiosqlite==0.20.0
python-dotenv==1.0.1
